requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
WTForms==3.1.1
email-validator==2.1.0.post1
bleach==6.1.0
redis==5.0.1
itsdangerous==2.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
websocket-client==1.6.4